backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
groq==0.11.0
python-multipart==0.0.6
cors==1.0.1
